Updated specification from 5 March 2024
An updated specification for FlexLink. The format is Book Chapter of a technical book. There is explanatory material, reference designs, and suggested implementations. A specification extraction will likely be necessary in the future in order to have a stand-alone specification document. Since the document is actively being edited, the extraction will happen after the book writing process.